requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.20.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
kombu==5.5.4
lxml==5.4.0
oauth2client==4.1.3
packaging==25.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
requests==2.32.4
rsa==4.9.1
simplegmail==4.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
